Added startup features
Now you can change the statup options, by clicking on the tray icon and
enabling/disabling the 'Launch on startup' option. Also, on the when you
launch Screen Cropper for the first time on any user, you will be asked
if you want  Screen Cropper to be run on Windows startup.